backend/requirements.txt
django==3.2.3
djangorestframework==3.12.4
djoser==2.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
flake8==7.0.0
flake8-isort==6.1.1
django-filter==23.5
django-extra-fields==3.0.2